Properly configure serializer/encoder in AvroMessageDecoderTest.
'CLIENTS-586 : ability to turn off register during serialize' was recently added to the schema registry code. It adds a flag
that disables registration of schemas. Because that flag is a boolean, if it is not initialized it defaults to false,
disabling registration. This fixes the AvroMessageDecoderTest to properly configure the serializer/encoder to avoid this
problem.