CC-1350: Corrected AK core’s missing dependency on SLF4J API
Recent changes in the 1.1.0-SNAPSHOT version of AK added code that depends on the SLF4J API library, but AK is not currently including that in the `org.apache:kafka_2.11` dependencies (see KAFKA-6313). This commit temporarily adds them via the `org.apache:kafka-clients` library, which does depend on the correct version of SLF4J API. This corrects the problem, but we’ll want to revert this when KAFKA-6313 is fixed.